#30514055 Завел отдельный сервис для распределения задач по статусам.
